refactor: removed EventDispatcherInternal and update docs
