+++ src/search/worker/file-Indexing-Worker.js
@@ -24,7 +24,7 @@
 const urlParams = new URLSearchParams(location.search);
 const debugMode = (urlParams.get('debug') === 'true');
 importScripts('../../phoenix/virtualfs.js');
-importScripts('../../utils/EventDispatcherInternal.js');
+importScripts('../../utils/EventDispatcher.js');
 importScripts('./search.js');
 
 virtualfs.debugMode = debugMode;
